error handling on comments DELETE

+++ app/api/comment_routes.py
@@ -10,7 +10,6 @@ def edit_comment(comment_id):
     #TODO check if user logged in is comment author
 
     comment_form = CommentForm()
-
     comment_form['csrf_token'].data = request.cookies['csrf_token']
     if comment_form.validate_on_submit():
         comment = Comment.query.get(comment_id)
@@ -25,10 +24,12 @@ def edit_comment(comment_id):
 @comment_routes.route("/<comment_id>", methods=["DELETE"])
 def delete_item(comment_id):
     #TODO user auth for deletion
-
     comment = Comment.query.get(comment_id)
+    if current_user.is_authenticated and current_user.get_id() == comment.user_id:
 
-    db.session.delete(comment)
-    db.session.commit()
+        db.session.delete(comment)
+        db.session.commit()
 
-    return "Successfully Deleted"
+        return "Successfully Deleted"
+    else:
+        return "User not logged in"
